requirements.txt
pandas==2.2.3
numpy==2.1.3
pyyaml==6.0.2
jinjasql==0.1.8
joblib==1.4.2
sqlalchemy==2.0.36
psycopg2==2.9.10

